update util file
